+++ resources/leiningen/new/luminus/keeframe/src/cljs/core.cljs
@@ -5,7 +5,6 @@
             [reagent.core :as r]
             [re-frame.core :as rf]
             [<<project-ns>>.ajax :refer [load-interceptors!]]
-            [<<project-ns>>.effects :as effects]
             [<<project-ns>>.routing :as routing])
   (:import goog.History))
 
@@ -82,6 +81,16 @@
 
 ;; -------------------------
 ;; Initialize app
+(def chain-links
+  [{;; Is the effect in the map?
+    :effect-present? (fn [effects] (:http effects))
+    ;; The dispatch set for this effect in the map returned from the event handler
+    :get-dispatch    (fn [effects]
+                       (get-in effects [:http :success-event]))
+    ;; Framework will call this function to insert inferred dispatch to next handler in chain
+    :set-dispatch    (fn [effects dispatch]
+                       (assoc-in effects [:http :success-event] dispatch))}])
+
 (defn mount-components []
   (rf/clear-subscription-cache!)
   (r/render [#'root-component] (.getElementById js/document "app")))
@@ -89,7 +98,7 @@
 (defn init! []
   (kf/start! {:debug?         true
               :router         (routing/->ReititRouter routing/router)
-              :chain-links    effects/chain-links
+              :chain-links    chain-links
               :initial-db     {}
               :root-component [root-component]})
   (load-interceptors!)