+++ src/leiningen/new/re_frame.clj
@@ -2,12 +2,16 @@
   (:require [leiningen.new.common :refer :all]))
 
 (def re-frame-assets
-  [["{{client-path}}/{{sanitized}}/core.cljs" "reframe/src/cljs/core.cljs"]
+  [["{{client-path}}/{{sanitized}}/ajax.cljs" "reframe/src/cljs/ajax.cljs"]
+   ["{{client-path}}/{{sanitized}}/core.cljs" "reframe/src/cljs/core.cljs"]
    ["{{client-path}}/{{sanitized}}/events.cljs" "reframe/src/cljs/events.cljs"]])
 
 (defn re-frame-features [[assets options :as state]]
   (if (some #{"+re-frame"} (:features options))
-    [(into (remove-conflicting-assets assets "core.cljs") re-frame-assets)
+    [(into (-> assets
+               (remove-conflicting-assets "ajax.cljs")
+               (remove-conflicting-assets "core.cljs"))
+           re-frame-assets)
      (-> options
          (assoc :re-frame true)
          (append-options :dependencies [['re-frame "0.10.5"]])